fix: removed unused function
