+++ pyeudiw/jwk/parse.py
@@ -1,21 +1,6 @@
-import cryptojwt
-import cryptojwt.jwk
 from cryptojwt.jwk.rsa import RSAKey, import_rsa_key
 
 from pyeudiw.jwk import JWK
-
-
-def adapt_key_to_JWK(key: dict | JWK | cryptojwt.jwk.JWK) -> JWK:
-    """Function adapt_key_to_JWK normalize key representation format to
-    the internal JWK.
-    """
-    if isinstance(key, JWK):
-        return key
-    if isinstance(key, dict):
-        return JWK(key)
-    if isinstance(key, cryptojwt.jwk.JWK):
-        return JWK(key.to_dict())
-    raise ValueError(f"not a valid or supported key format: {type(key)}")
 
 
 def parse_key_from_x5c(x5c: list[str]) -> JWK:
